Also use actor_type to determine if an account is a bot in antiFollowbotPolicy
An account can be marked as "Bot". Those accounts will have actor_type="Service". Bots who follow someone out of the blue should be considered a followbot. However, there may be bots that need to follow their target. In that case the target needs to be able to show it's ok to be followed. To do this, we check if the target follows the bot.