Add test "Pleroma:report is hidden for non-superusers"
I added logic in mastodon_api for this and it works, but it should be done better.

I now use `User.all_superusers() |> Enum.any?(fn u -> u.id == user.id end)` to see if someone is a superuser.
This should really just be `User.superuser?(user)`.

The problem is that it somehow doesn't work. The `user` we get in `get_notifications` is supposedly non moderator, while the same user returned with
`User.all_superusers()` is moderator. So why does the same user have different parameters when fetched differently?