Always provide pleroma.deactivated in masto_api account_view
When a user is deactivated, "normal" users can't see them. Moderators and admins can.

There was logic that only added the "deactivated" key for admins, meaning that mods were able to see the account, but not that it was deactivated.

I didn't really see a reason to not just give that info always, so I removed that part of the logic.

Now "normal" users will still not see the account. Any one who can see the account will see if they are deactivated or not.